12.14 1：用例页面增删改查同时处理waits表

+++ model/baseModel.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type BaseModel struct {
-	ID        int64      `gorm:"column:id;primary_key" json:"id"`
+	ID        int        `gorm:"column:id;primary_key" json:"id"`
 	CreatedAt time.Time  `gorm:"column:created_at" json:"created_at"`
 	UpdatedAt time.Time  `gorm:"column:updated_at" sql:"index" json:"updated_at"`
 	DeletedAt *time.Time `gorm:"column:deleted_at" sql:"index" json:"deleted_at"`
